Renaming enrichment_tools to enrichment_result. This files stores the EnrichmentResult class.
Removing unused notebook_tools.
