+++ magine/enrichment/__init__.py
@@ -1,4 +1,4 @@
+from .enrichment_result import load_enrichment_csv
 from .enrichr import Enrichr
-from .tools import load_enrichment_csv
 
 __all__ = ['load_enrichment_csv', 'Enrichr']
